+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 ### New Features
 
+- Watch mode now works with folders, reacting to source-code changes and adding templates/source files.
+- When using watch mode, status info will be displayed in the generated code so that you don't need to look at console at all.
 - You can now access type's `methods` and `initializers`
 
 ## 0.4.8